refactor: RESTful 규칙에 맞게 엔드포인트 수정 #75 (#76)
* refactor: RESTful 규칙에 맞게 엔드포인트 수정 #75

* refactor: Principal 객체에서 memberId 추출하여 사용하도록 변경 #75

* refactor: endpoint 변경 #75
