+++ src/main/java/com/market/bucket/controller/BucketController.java
@@ -1,11 +1,11 @@
 package com.market.bucket.controller;
 
-
 import com.market.bucket.dto.request.BucketSaveRequest;
 import com.market.bucket.dto.response.BucketDiscriminationResponse;
 import com.market.bucket.dto.response.BucketProductResponse;
 import com.market.bucket.service.BucketService;
 import com.market.core.response.BfResponse;
+import com.market.core.security.principal.PrincipalDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -22,7 +22,7 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/bucket")
 @RequiredArgsConstructor
-@Tag(name = "장바구니 API", description = "장바구니 관련 API 입니다.")
+@Tag(name = "장바구니", description = "장바구니 관련 API")
 public class BucketController {
 
     private final BucketService bucketService;
@@ -32,20 +32,20 @@ public class BucketController {
             @ApiResponse(responseCode = "200", description = "가게 상세 조회 성공")
     })
     @GetMapping
-    public ResponseEntity<BfResponse<BucketProductResponse>> findBucket(@AuthenticationPrincipal String memberId) {
-        return ResponseEntity.ok(new BfResponse<>(bucketService.findBucket(Long.parseLong(memberId))));
+    public ResponseEntity<BfResponse<BucketProductResponse>> findBucket(@AuthenticationPrincipal PrincipalDetails principalDetails) {
+        return ResponseEntity.ok(new BfResponse<>(bucketService.findBucket(Long.parseLong(principalDetails.getUsername()))));
     }
 
-    @Operation(summary = "장바구니 다른 가게 상품 여부 판단", description = "장바구니에 다른 가게 상품이 존재하는지 여부 판단입니다.")
+    @Operation(summary = "장바구니 다른 가게 상품 여부 확인", description = "장바구니에 현재 가게와 다른 가게의 상품이 존재하는지 여부를 확인합니다..")
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "판단 성공")
+            @ApiResponse(responseCode = "200", description = "가게 상품 여부 확인 성공")
     })
-    @GetMapping("/discrimination/{marketId}")
+    @GetMapping("/{marketId}/verification/product")
     public ResponseEntity<BfResponse<BucketDiscriminationResponse>> discriminateBucket(
-            @AuthenticationPrincipal String memberId,
+            @AuthenticationPrincipal PrincipalDetails principalDetails,
             @Parameter(description = "현재 장바구니에 담고자 하는 상품의 가게 ID 입니다.") @PathVariable("marketId") Long marketId) {
         return ResponseEntity.ok(new BfResponse<>(new BucketDiscriminationResponse(
-                bucketService.discriminateBucket(marketId, Long.parseLong(memberId)))));
+                bucketService.discriminateBucket(marketId, Long.parseLong(principalDetails.getUsername())))));
     }
 
     @Operation(summary = "장바구니 상품 추가", description = "장바구니 상품 추가 요청입니다.")
@@ -54,7 +54,6 @@ public class BucketController {
     })
     @PostMapping("/{marketId}")
     public ResponseEntity<BfResponse<String>> saveBucket(
-            @AuthenticationPrincipal String memberId,
             @RequestBody BucketSaveRequest bucketSaveRequest,
             @Parameter(description = "현재 장바구니에 담고자 하는 상품의 가게 ID 입니다.") @PathVariable("marketId") Long marketId) {
         bucketService.saveBucket(Long.parseLong("1"), marketId, bucketSaveRequest.getProducts());
